Complete Phase 5.1: GGUF v3 test data regression fix (457 tests passing)

+++ llm-runner/src/transformer/tokenizer.rs
@@ -231,7 +231,7 @@ mod tests {
 
         for kv in &kv_pairs {
             let key_bytes = kv.key.as_bytes();
-            buf.extend_from_slice(&(key_bytes.len() as u64).to_le_bytes());
+            buf.extend_from_slice(&(key_bytes.len() as u32).to_le_bytes());
             buf.extend_from_slice(key_bytes);
             buf.extend_from_slice(&kv.value_type.to_u32().to_le_bytes());
             write_kv_value(&mut buf, &kv.value);
@@ -416,7 +416,7 @@ mod tests {
         buf.extend_from_slice(&(kv_pairs.len() as u64).to_le_bytes());
         for kv in &kv_pairs {
             let key_bytes = kv.key.as_bytes();
-            buf.extend_from_slice(&(key_bytes.len() as u64).to_le_bytes());
+            buf.extend_from_slice(&(key_bytes.len() as u32).to_le_bytes());
             buf.extend_from_slice(key_bytes);
             buf.extend_from_slice(&kv.value_type.to_u32().to_le_bytes());
             write_kv_value(&mut buf, &kv.value);
@@ -478,7 +478,7 @@ mod tests {
         buf.extend_from_slice(&(kv_pairs.len() as u64).to_le_bytes());
         for kv in &kv_pairs {
             let key_bytes = kv.key.as_bytes();
-            buf.extend_from_slice(&(key_bytes.len() as u64).to_le_bytes());
+            buf.extend_from_slice(&(key_bytes.len() as u32).to_le_bytes());
             buf.extend_from_slice(key_bytes);
             buf.extend_from_slice(&kv.value_type.to_u32().to_le_bytes());
             write_kv_value(&mut buf, &kv.value);
